workspace/tumor_segmentation/requirements.txt
torch>=2.1
torchvision>=0.16
monai>=1.3
numpy>=1.24
nibabel>=5.0
pydicom>=2.4
Pillow>=10.0
